messed up menu
I changed the blockquote in my about page to be col-lg so it was centered, then all my menu items stacked upon one another (and the ugly bullets appeared too). I've noticed it's also appeared on the home page that I'm working on now too. Finally, I've been trying to change the background of the active menu item, but I haven't figured it out. One thing I'm about to try is maybe editing the CSS (background-color) for an active link to see if that does it.